refactor - modularised the nav bar component
